requirements.txt
pyshark>=0.4
xgboost>=1.7
scikit-learn>=1.0
pandas>=1.5
numpy>=1.21
flask>=2.0
twilio>=7.0
requests>=2.28
reportlab>=4.0
pyyaml>=6.0
python-dotenv>=0.19
